Feat: separate functions and make it more clear
Refactor code
